+++ src/main/java/org/j2overhead/m1ke/model/Repository.java
@@ -1,7 +1,26 @@
 package org.j2overhead.m1ke.model;
 
+import java.nio.file.Path;
 import java.util.ArrayList;
 
 public class Repository {
+    // path to m1ke repository
+    Path path;
     ArrayList<Branch> branches;
+
+    public Path getPath() {
+        return path;
+    }
+
+    public void setPath(Path path) {
+        this.path = path;
+    }
+
+    public ArrayList<Branch> getBranches() {
+        return branches;
+    }
+
+    public void setBranches(ArrayList<Branch> branches) {
+        this.branches = branches;
+    }
 }
