+++ src/main/java/org/j2overhead/m1ke/service/BranchServiceImpl.java
@@ -1,4 +1,4 @@
-package org.j2overhead.m1ke;
+package org.j2overhead.m1ke.service;
 
 import org.j2overhead.m1ke.model.Branch;
